+++ crates/full-node/sov-sequencer/src/preferred/mod.rs
@@ -6,11 +6,12 @@ mod block_executor;
 mod cache_warm_up_executor;
 mod db;
 mod executor_events;
-mod inner;
+//mod inner;
 mod preferred_blob_sender;
 mod replica;
 mod side_effects;
 mod state_root_compute;
+mod sync_sequencer_state;
 mod transaction_subscriptions;
 mod update_state;
 
@@ -25,7 +26,6 @@ use db::rocksdb::RocksDbBackend;
 use db::{PreferredSequencerDb, PreferredSequencerReadBatch, PreferredSequencerReadBlob};
 pub use full_node_configs::sequencer::{PreferredSequencerConfig, RecoveryStrategy};
 use futures::Stream;
-use inner::*;
 use preferred_blob_sender::PreferredBlobSender;
 use serde_with::serde_as;
 use side_effects::SideEffectsTask;
@@ -54,6 +54,7 @@ use std::pin::Pin;
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
+use sync_sequencer_state::*;
 use tokio::sync::mpsc::{self};
 use tokio::sync::{broadcast, watch};
 use tokio::task::JoinHandle;
