+++ crates/full-node/sov-sequencer/src/preferred/sync_sequencer_state/mod.rs
@@ -1,0 +1,230 @@
+use super::batch_size_tracker::BatchSizeTracker;
+use crate::preferred::block_executor::RollupBlockExecutor;
+use crate::preferred::cache_warm_up_executor::CacheWarmUpExecutor;
+use crate::preferred::db::BatchToStore;
+use crate::preferred::executor_events::ExecutorEventsSender;
+use crate::preferred::replica::event_handler::ReplicaError;
+use crate::preferred::AcceptedTx;
+use crate::preferred::BatchCreationError;
+use crate::preferred::Confirmation;
+use crate::preferred::DbEvent;
+use crate::preferred::PreferredSeqOperation;
+use crate::preferred::RollupBlockExecutorConfig;
+use crate::preferred::TxResultWriter;
+use crate::{SequencerConfig, SequencerNotReadyDetails, TxHash};
+use full_node_configs::sequencer::PreferredSequencerConfig;
+pub(crate) use inner::*;
+use sov_blob_sender::BlobInternalId;
+use sov_blob_storage::SequenceNumber;
+use sov_db::ledger_db::LedgerDb;
+use sov_modules_api::capabilities::RollupHeight;
+use sov_modules_api::{FullyBakedTx, Runtime, Spec, StateUpdateInfo};
+use sov_state::Storage;
+use std::collections::BTreeMap;
+use std::sync::atomic::{AtomicU32, AtomicU64, AtomicUsize};
+use std::sync::Arc;
+pub(crate) use sync_state::*;
+use tokio::sync::{mpsc, oneshot, watch};
+pub(crate) use updator::*;
+
+mod inner;
+mod sync_state;
+mod updator;
+
+const CHANNEL_SIZE: usize = 128;
+
+enum Message<S: Spec, Rt: Runtime<S>> {
+    NextSequenceNumber {
+        resp: oneshot::Sender<SequenceNumber>,
+        reason: &'static str,
+    },
+    FetchCompletedBatches {
+        resp: oneshot::Sender<FetchBatches>,
+        next_sequence_number: u64,
+        reason: &'static str,
+    },
+    SequencerConditions {
+        resp: oneshot::Sender<PreferredSeqOperation<S, Rt>>,
+        info: StateUpdateInfo<S::Storage>,
+        next_sequence_number_according_to_node: u64,
+        reason: &'static str,
+    },
+    CheckReadiness {
+        resp: oneshot::Sender<Result<(), SequencerNotReadyDetails>>,
+        max_concurrent_blobs: usize,
+        height_to_stop_at: Option<RollupHeight>,
+        reason: &'static str,
+    },
+
+    AcceptTx {
+        resp: oneshot::Sender<AcceptTxRet<S, Rt>>,
+        baked_tx: FullyBakedTx,
+        tx_hash: TxHash,
+        original_tx_queue_id: u64,
+        reason: &'static str,
+    },
+
+    FinalCatchup {
+        resp: oneshot::Sender<anyhow::Result<ProcessFinalCatchupData>>,
+        info: StateUpdateInfo<S::Storage>,
+        db_event_subscription: mpsc::Receiver<DbEvent>,
+        executor: Box<RollupBlockExecutor<S, Rt>>,
+        node_state_root: <S::Storage as Storage>::Root,
+        data: ProcessFinalCatchupData,
+        reason: &'static str,
+    },
+    PruneSequencerDb {
+        reason: &'static str,
+    },
+    ForceOverwriteStateForRecovery {
+        info: StateUpdateInfo<S::Storage>,
+        reason: &'static str,
+    },
+    WaitNodeResync {
+        info: StateUpdateInfo<S::Storage>,
+        distance: u64,
+        reason: &'static str,
+    },
+    #[cfg(feature = "test-utils")]
+    ForceCloseCurrentBatch {
+        reason: &'static str,
+    },
+    ProofBlob {
+        blob_id: BlobInternalId,
+        data: Arc<[u8]>,
+        reason: &'static str,
+    },
+    TriggerBatchProductionIfConvenient {
+        reason: &'static str,
+    },
+    SimpleStateUpdate {
+        info: StateUpdateInfo<S::Storage>,
+    },
+    ReplicaBatchStartMsg {
+        resp: oneshot::Sender<Result<(), ReplicaError<S>>>,
+        batch_from_master: BatchToStore,
+        reason: &'static str,
+    },
+    ReplicaNewTx {
+        resp: oneshot::Sender<Result<(), ReplicaError<S>>>,
+        seq_nr_from_master: u64,
+        tx_hash: TxHash,
+        baked_tx: FullyBakedTx,
+        reason: &'static str,
+    },
+    ReplicaCloseCurrentBatch {
+        resp: oneshot::Sender<Result<(), ReplicaError<S>>>,
+        batch_from_master: BatchToStore,
+        reason: &'static str,
+    },
+}
+
+impl<S: Spec, Rt: Runtime<S>> Message<S, Rt> {
+    fn priority(&self, runtime: &Rt) -> u64 {
+        // Computes the priority of the message. Note that the implementation of the queue assumes that transactions never have higher priority than internal
+        // sequencer messages - if we change this assumption, we'll need to update the implementation of sequencer state updater to handle this.
+        match self {
+            Message::AcceptTx { baked_tx, .. } => {
+                let priority = runtime.get_transaction_priority(baked_tx);
+                priority as u64
+            }
+            _ => u64::MAX,
+        }
+    }
+}
+
+pub(crate) fn create<S, Rt>(
+    api_ledger_db: LedgerDb,
+    latest_info: StateUpdateInfo<S::Storage>,
+    tx_queue_id: Arc<AtomicU64>,
+    batch_execution_time_limit_micros: u64,
+    seq_config: SequencerConfig<S::Address, PreferredSequencerConfig>,
+    shutdown_receiver: watch::Receiver<()>,
+    shutdown_sender: watch::Sender<()>,
+    executor_events_sender: ExecutorEventsSender<S, Rt>,
+    sequence_number_of_next_blob: SequenceNumber,
+    in_flight_blobs: Arc<AtomicUsize>,
+    stop_at_rollup_height: Option<RollupHeight>,
+    rollup_exec_config: RollupBlockExecutorConfig<S>,
+    tx_cache_writer: TxResultWriter<S, Rt>,
+    cache_warm_up_executor: CacheWarmUpExecutor<S>,
+) -> (
+    SynchronizedSequencerState<S, Rt>,
+    SequencerStateUpdator<S, Rt>,
+)
+where
+    S: Spec,
+    Rt: Runtime<S>,
+{
+    let (message_sender, message_receiver) = mpsc::channel(CHANNEL_SIZE);
+
+    let is_ready = if seq_config.sequencer_kind_config.is_replica {
+        Err(SequencerNotReadyDetails::ReplicaMode)
+    } else {
+        Err(SequencerNotReadyDetails::Startup)
+    };
+
+    let inner = Inner {
+        api_ledger_db,
+        executor: RollupBlockExecutor::new(
+            &latest_info,
+            rollup_exec_config.clone(),
+            seq_config.clone(),
+            Default::default(),
+        ),
+        latest_info,
+        tx_queue_id,
+        batch_execution_time_limit_micros,
+        batch_size_tracker: BatchSizeTracker::new(seq_config.max_batch_size_bytes),
+        seq_config: seq_config.clone(),
+        shutdown_receiver: shutdown_receiver.clone(),
+        shutdown_sender,
+        executor_events_sender,
+        sequence_number_of_next_blob,
+        in_flight_blobs,
+        has_finished_startup: false,
+        metrics: Vec::with_capacity(128),
+        is_ready,
+        stop_at_rollup_height,
+        rollup_exec_config,
+        tx_cache_writer,
+        cache_warm_up_executor,
+    };
+
+    let channel_size = Arc::new(AtomicU32::new(0));
+    (
+        SynchronizedSequencerState {
+            inner,
+            channel_size: channel_size.clone(),
+            message_receiver,
+            heap: BTreeMap::new(),
+            runtime: Default::default(),
+        },
+        SequencerStateUpdator {
+            message_sender,
+            channel_size,
+            shutdown_receiver,
+        },
+    )
+}
+
+type AcceptTxRet<S, Rt> =
+    Result<oneshot::Receiver<AcceptedTx<Confirmation<S, Rt>>>, AcceptTxError<S>>;
+
+#[derive(Debug)]
+pub(crate) enum AcceptTxError<S: Spec> {
+    SequencerOverloaded503,
+    NotFullySynced(SequencerNotReadyDetails),
+    BatchError {
+        batch_creation_error: BatchCreationError,
+        nb_of_concurrent_blob_submissions: usize,
+    },
+    NewTxError(DoNewTxError<S>),
+}
+
+#[derive(Debug)]
+pub(crate) struct ProcessFinalCatchupData {
+    pub(crate) batches_count: u64,
+    pub(crate) transactions_count: usize,
+    pub(crate) batch_is_in_progress: bool,
+}
